- Changed project into a maven project with an additional zip assembly. - Added dependencies for using the "addon-test-helpers" addon which makes (manual) testing much easier. More on that in later commits.
