--- WebContent/META-INF/MANIFEST.MF
@@ -1,8 +1,0 @@
-Manifest-Version: 1.0
-Vaadin-License-Title: Apache License 2.0
-Implementation-Title: NumberField (Vaadin 7)
-Implementation-Version: 0.1.6
-Vaadin-Widgetsets: org.vaadin.ui.NumberFieldWidgetset
-Vaadin-Package-Version: 1
-Class-Path: 
-
