+++ assembly/MANIFEST.MF
@@ -1,0 +1,7 @@
+Manifest-Version: 1.0
+Vaadin-Package-Version: 1
+Vaadin-Addon: ${Vaadin-Addon}
+Vaadin-License-Title: ${Vaadin-License-Title}
+Vaadin-Widgetsets: ${Vaadin-Widgetsets}
+Implementation-Title: ${Implementation-Title}
+Implementation-Version: ${Implementation-Version}
