+++ src/main/java/org/vaadin/ui/shared/numberfield/Constants.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2012 essendi it GmbH
+ * Copyright 2014-2016 Vaadin Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * 		http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package org.vaadin.ui.shared.numberfield;
+
+public interface Constants {
+
+    static final String CSS_CLASSNAME = "numberfield";
+    static final String CSS_CLASSNAME_ON_ERROR = "numberfield-error";
+
+    static final String ATTRIBUTE_SERVER_FORMATTED_VALUE = "formattedValue";
+
+    static final String NEGATIVE_PREFIX = "-";
+
+}
