+++ src/main/resources/org/vaadin/ui/error_underline.png
