+++ src/test/webapp/VAADIN/themes/numberfieldtest/styles.scss
@@ -1,5 +1,5 @@
-@import "addons.scss";
-@import "numberfieldtest.scss";
+@import "addons";
+@import "numberfieldtest";
 
 /* This file prefixes all rules with the theme name to avoid causing conflicts with other themes. */
 /* The actual styles should be defined in numberfieldtest.scss */
